pyproject.toml
[project]
name = "ecommerce-ad-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "joblib>=1.5.3",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "lxml>=6.0.2",
    "mlflow>=1.27.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.19.0",
    "uvicorn>=0.42.0",
    "xgboost>=3.2.0",
]
